acomodo funciones de inputs en archivos + renombro variables + comentarios
